Extract FakeNuGetPackageCache to shared test utility
